extract ReturnValue into its own file
